+++ apps/shadcnpreset/components/preset-swatch/components/cards/observability-card.tsx
