+++ apps/shadcnpreset/components/preset-swatch/components/color-specimen.tsx
